+++ pyi.py
@@ -643,6 +643,28 @@ class PyiVisitor(ast.NodeVisitor):
                 self.error(statement, Y013)
 
     def visit_FunctionDef(self, node: ast.FunctionDef) -> None:
+        # Raise an error for defining __str__ or __repr__ on a class, but only if:
+        # 1). The method is not decorated with @abstractmethod
+        # 2). The method has the exact same signature as object.__str__/object.__repr__
+        if (
+            self.in_class
+            and node.name in {"__repr__", "__str__"}
+            and isinstance(node.returns, ast.Name)
+            and node.returns.id == "str"
+            and not any(
+                isinstance(deco, ast.Name) and deco.id == "abstractmethod"
+                for deco in node.decorator_list
+            )
+        ):
+            all_args = node.args
+            # pos-only args don't exist on 3.7
+            pos_only_args: list[ast.arg] = getattr(all_args, "posonlyargs", [])
+            pos_or_kwd_args = all_args.args
+            kwd_only_args = all_args.kwonlyargs
+
+            if ((len(pos_only_args) + len(pos_or_kwd_args)) == 1) and not kwd_only_args:
+                self.error(node, Y029)
+
         self._visit_function(node)
 
     def visit_AsyncFunctionDef(self, node: ast.AsyncFunctionDef) -> None:
@@ -882,3 +904,4 @@ Y025 = (
 Y026 = "Y026 Use typing_extensions.TypeAlias for type aliases"
 Y027 = 'Y027 Use {good_cls_name} instead of "{bad_cls_alias}"'
 Y028 = "Y028 Use class-based syntax for NamedTuples"
+Y029 = "Y029 Defining __repr__ or __str__ in a stub is almost always redundant"